MGMT-8149: Add OPENSHIFT_VERSIONS template parameter back (#2877) (#2878)
This was removed without removing the env var from the deployment and
also before we stopped setting it in the SaaS deployment configuration
(saas.yaml).

Add it back for now and we'll remove it properly after cleaning up the
other deployment configs.

Currently not having this in the template is causing our deploys to fail
with:
`time="2021-11-04T12:26:13Z" level=fatal msg="Failed to parse supported openshift versions JSON ${OPENSHIFT_VERSIONS}" func=main.main.func1 file="/go/src/github.com/openshift/origin/cmd/main.go:196" error="invalid character '$' looking for beginning of value"`
This is because we're writing the literal string `${OPENSHIFT_VERSIONS}`
as the env var because the template doesn't recognize OPENSHIFT_VERSIONS
as a parameter that requires substitution.

Co-authored-by: Nick Carboni <ncarboni@redhat.com>