MGMT-8074: Make the assisted-service more dry run friendly (#2822)
Dry run modes have been introduced to the agent/installer (and in the
future also to the controller):

https://github.com/openshift/assisted-installer-agent/pull/250
https://github.com/openshift/assisted-installer/pull/371

To make these dry run modes work more easily, the service required some
minor modifcations, implemented by this commit.

The changes are as follows -
1) Propagate host agent environment to next-step-runner and inventory
This commit makes it so the `CONTAINERS_CONF` and `CONTAINERS_STORAGE_CONF` environment
variables from the host agent also propagate to the the next_step_runner container.

This allows the assisted-swarm to freely control how other environment variables
propagate to the step containers, and also how podman storage is
managed for those containers.

Currently CONTAINERS_CONF is used to tell podman to automatically relay
certain environment variables to the step containers (with the
.containers.env containers.conf TOML configuration).

CONTAINERS_STORAGE_CONF is used to configure podman storage in such a
way that swarm agents don't share the same container storage, but do
share image storage and overlayfs layers. This helps with performance.

2) mtab copy file path - when launching the inventory command, we tell
the agent to copy the `/etc/mtab` file to `/root/mtab`. This caused an
issue when running multiple agents on the same host (using the dry run
mode), because they were fighting over the `/root/mtab` host path. This
commit appends the host ID to the `/root/mtab` path to make sure that
each swarm host will work with a different file and the will not
fight over the same path.

3) Use the new installer `--format-disk` flag rather than use plain `dd`
commands directly. This gives the installer more flexibility in deciding
whether to format the disks or just to pretend to format them as it does
in dry mode.

This commit should NOT be merged before
https://github.com/openshift/assisted-installer/pull/371 is merged,
because only in that PR the `--format-disk` commandline flags are
implemented.