Bug 2011745: Document ignition versions (#2879)
The fact that the discovery ignition and host pointer ignitions can use
different versions was confusing to some users, but is generally
unavoidable as the host ignition version will vary with what OCP version
is being installed.

This adds some documentation which should help clear up what versions we
support for each endpoint.

https://bugzilla.redhat.com/show_bug.cgi?id=2011745