MGMT-7254: Fix subsystem tests after adding image service
Specifically this uses the cluster/infraEnv image download url to get
the iso rather than using the api client. When using the image service
that endpoint will always fail as we don't upload the images to s3.

This also removes some tests that don't make sense when using the image
service like checking that an image is no longer available. And also
adds an infra-env event when we set the download url to the image service
like we did previously.

Lastly it changes the behavior of the assisted-service-iso endpoints to
just always return a 400 if we're using the image service.
These endpoints require that we have the rhcos images uploaded to s3 and
that won't be the case when the image service is in use.