MGMT-8192: Logs_cmd command fails if insecure flag is set to true (#2902)
When insecure=true logs sender will try to use certificate that was set in --cacert flag.
Service currently always sets "cacert" flag with path that was not mounted and not exist.

This commit fixes setting cacert flag all the time and will set it only
in case certificate path was provided, same as in install_cmd