MGMT-7897: Add Image-Token authentication (#2824)
* MGMT-7897: add the image token key to the infra env

Adds a new column to the infraEnv (`image_token_key`) which is used to
store the encryption key for the auth tokens generated for each
infraEnv.

This commit also generates a new token when a new infraEnv is created.

* MGMT-7897: add functions to create jwts for a symmetric key

* MGMT-7897: add new auth type and generate from swagger

* MGMT-7897: add image token auth implementation

Image auth is only implemented for rhsso and is accomplished in two
parts. Authentication parses the token and validates that the signature
is valid using the token for the infraEnv in the "sub" claim as well
as ensures that the token is not expired (using the "exp" claim).
Authorization then ensures that the infraEnv in the "sub" claim is the
same as the one in the request path.

* MGMT-7897: sign the image download url with the infraEnv image token

* MGMT-7897: use http package constants in local authenticator

* MGMT-7897: Add infra env token key regenerate endpoint

Calling this endpoint will generate a new signing key for the infraEnv.
This will effectively invalidate all existing unexpired tokens and
pre-signed urls.

* MGMT-7897: Add migration to add image token key to existing infraEnvs

When this is deployed to the SaaS we should always have a value for
image_token_key.