MGMT-8205: Add endpoint to regenerate infraEnv download URL (#2919)
* MGMT-8205: Add endpoint for requesting a new infraEnv image URL

* MGMT-8205: Extract logic for creating a download url

This will allow us to share this logic in all the following cases where
we need to recreate the URL:

- infraEnv is updated (current use case)
- new image token key is created
- new signed url is requested

* MGMT-8205: Add an API for regenerating the infraEnv download url

This endpoint returns a new download url for a discovery image
associated with the infraEnv when the image service is in use.

For local and rhsso auth types this will generate a new auth token and
embed it into the returned url.

It also saves the generated url to the infraEnv, but this is for
compatibility purposes and may change in the future.

`GET /v2/infra-envs/{infra_env_id}/downloads/image-url`
Example response:
`{"url": "https://image-service.example.com/images/{infra_env_id}?image_token=<some-jwt>&version=4.9&arch=x86_64&type=full-iso", "expires_at": "2021-11-15T21:39:51.000Z"}`

* MGMT-8205: Return bad request (400) if we're not using the image service

We don't want to overwrite any existing download urls if we're not using
the image service. So at least for now just error the endpoint if we
don't have an image service url configured.

* MGMT-8205: Use the correct version in the url

Also, the similar code that was looking for this version in
updateExternalImageInfo was checking for the Version field, but then
using the OpenshiftVersion field. This commit also fixes that.

* MGMT-8205: Update infra-env expires_at property

* MGMT-8205: update enhancement

This makes the response from the new presigned url endpoint agree with
what was implemented