MGMT-7254: fix kubeapi subsystem tests
Add the connection file step to the static network config validation
function.

This also means that the error will be reported differently so we match
a substring instead of the exact error. Additionally, the infra-env
never gets created because the original validation fails, so we can't
check the generated flag