MGMT-7254: Ensure the configmap has routable base urls
Prior to this commit we would almost always get some variety of
`<pending>` for both the `SERVICE_BASE_URL` and `IMAGE_SERVICE_BASE_URL`
env vars.

This wasn't a problem in our end to end tests so far because we always
replace the `SERVICE_BASE_URL` as a part of the test-infra automation,
but now that we've added the `IMAGE_SERVICE_BASE_URL` we need to ensure
that it's set properly using just the deploy code in this repo.