MGMT-7716: Add IMAGE_SERVICE_BASE_URL to template (#2983)
This allows us to specify this parameter in the saas file when deploying
to various environments in the cloud.

It is not marked as required because, for now, it will only be enabled
in integration. The default value for this was always an empty string in
our source so this won't change the behavior of stage or prod.