MGMT-8147: Expose OlmOperators through KubeApi
Assisted service already supports enabling some olm operators that would be deployed as part of the day1 install
process. This is currently only exposed through the HTTP API. We need to also expose them through the KubeAPI for
consistency and because we want to be able to install operators like ODF and OCS as part of the day1 step on specific
scenarios (ZTP factory pipeline).

Signed-off-by: Flavio Percoco <flavio@redhat.com>