MGMT-8372: Removing data from "Custom manifests" usage (#2997)
In each `feature usage` there is an option to add additional data related to that feature.
`Custom manifests` feature was storing a map of filenames in that `data` field.

We have no need for the file names at the moment, and the data size is significantly bigger.

This PR will no longer store that data.