MGMT-7254: Deploy and use image service for subsystem (#2884)
* MGMT-7254: Deploy and use image service for subsystem

This commit adds the image service deployment spec, script, and make
target. It also configures assisted service to direct image downloads to
the image service rather than generating the images and uploading them
to s3

* MGMT-7254: Also wait for the image service when deploying

* MGMT-7254: Fix subsystem tests after adding image service

Specifically this uses the cluster/infraEnv image download url to get
the iso rather than using the api client. When using the image service
that endpoint will always fail as we don't upload the images to s3.

This also removes some tests that don't make sense when using the image
service like checking that an image is no longer available. And also
adds an infra-env event when we set the download url to the image service
like we did previously.

Lastly it changes the behavior of the assisted-service-iso endpoints to
just always return a 400 if we're using the image service.
These endpoints require that we have the rhcos images uploaded to s3 and
that won't be the case when the image service is in use.

* MGMT-7254: fix kubeapi subsystem tests

Add the connection file step to the static network config validation
function.

This also means that the error will be reported differently so we match
a substring instead of the exact error. Additionally, the infra-env
never gets created because the original validation fails, so we can't
check the generated flag

* MGMT-7254: log the image service url when we're using it

* MGMT-7254: Ensure the configmap has routable base urls

Prior to this commit we would almost always get some variety of
`<pending>` for both the `SERVICE_BASE_URL` and `IMAGE_SERVICE_BASE_URL`
env vars.

This wasn't a problem in our end to end tests so far because we always
replace the `SERVICE_BASE_URL` as a part of the test-infra automation,
but now that we've added the `IMAGE_SERVICE_BASE_URL` we need to ensure
that it's set properly using just the deploy code in this repo.

* MGMT-7254: Use path based routing for TARGET=oc-ingress

This allows access to both the image service and assisted service with
only the one ingress. In ocp this will create separate routes for each
path.