MGMT-8332: Disable machine network autogeneration for dual-stack
This PR changes the logic behind automatic calculation of Machine
Network so that it does not apply for dual-stack clusters. For those we
require user to explicitly provide machine networks according to
requirements of the OCP as well as OVN-K8s.

As a side effect this PR fixes a bug which made it impossible to
configure Machine Networks before cluster had Agents registered,
effectively requiring multiple API calls to form a correct dual-stack
cluster.

Contributes-to: MGMT-8300
Contributes-to: MGMT-8332