MGMT-8300: Improve behaviour of dual-stack via REST API (#3042)
* MGMT-8300: Allow configuring dual-stack networking in single step

This PR changes the logic of updateNonDhcpNetworkParams so that changing
cluster state between single- and dual-stack is possible in a single API
call.

Previously, because the check of the "stackness" of the cluster was
happening without taking the updated params into consideration, at least
2 API calls were needed to change the status of the cluster. With this
change in order to check what is the "stackness" of the cluster we are
using a merge of 2 states

* current cluster network configuration
* updated configuration passed as an update parameter

Because of this, we no longer use the stale state.

Contributes-to: MGMT-8300

* MGMT-8332: Disable machine network autogeneration for dual-stack

This PR changes the logic behind automatic calculation of Machine
Network so that it does not apply for dual-stack clusters. For those we
require user to explicitly provide machine networks according to
requirements of the OCP as well as OVN-K8s.

As a side effect this PR fixes a bug which made it impossible to
configure Machine Networks before cluster had Agents registered,
effectively requiring multiple API calls to form a correct dual-stack
cluster.

Contributes-to: MGMT-8300
Contributes-to: MGMT-8332