MGMT-7330: Make `ROOT_DIR`, `COVER_PROFILE` and `GINKGO_REPORTFILE` overridable for running unit tests in CI (#3038)
* MGMT-7330: Make `ROOT_DIR`, `COVER_PROFILE` and `GINKGO_REPORTFILE` overridable for running unit tests in CI
When running the unit tests in CI, the image is built as root but run as non-root.
CI isn't able to create the reports directory and results in the error
`mkdir: cannot create directory '/opt/app-root/src/reports': Permission denied`
During runtime, directories can't be added outside of `/tmp`.
To address this, `ROOT_DIR`, `COVER_PROFILE`, and `GINKGO_REPORTFILE` are made overridable. CI when running the unit tests will set these to the below values:
- ROOT_DIR                  "/tmp"
- COVER_PROFILE             "/tmp/unit_coverage.out"
- GINKGO_REPORTFILE         "/tmp/junit_unit_test.xml"

* MGMT-7330: Resolve `TEST_SCENARIO` value to run unit tests

* MGMT-7330: Fix coverage.xml path