MGMT-8448: Redirect v1 image downloads to the image service
This PR backports https://github.com/openshift/assisted-service/pull/3023
to redirect v1 downloads to image service. This is required to avoid failing
when calling the service directly for downloading the discovery image.

This scenario is relevant for the 'cloud_hotfix_releases' branch since the
deployment is actually done using 'master' branch - which means that the
assisted-image-service is being used.

Specifically, this change fixes day2-periodic test in CI as that e2e test
still uses v1 API.