MGMT-8047: Approve the CSRs after the installation
Day1 CSRs are approved by assisted controller.  Since the controller is
not aware of day2 workers, until now these CSRs needed to be approved
manually.
This modification, adds capability to assisted service, to approve these
CSRs.  In order for this to work, cluster kubeconfig has to be available in S3
(which is not always true for day2 hosts), and the cluster needs to be
available when using this kubeconfig.
There are two parts to this solution.
In host monitoring, each day2 host that is not in done stage, is checked against its corresponding node
(using kubeapi).  If the node is ready then the host is moved to done.
In case there is a problem acccessing the day1 cluster and the day2 host
is rebooting, it is moved to done (as it is today).  In this case it is
the user responsibility to approve the CSRs.
In cluster monitoring, if there is a day2 cluster with hosts that are
not in done stage yet, it tries to approve all the CSRs using kubeapi.
For the usage of these monitoring loops, a package with the needed
kubeapi operations was created.