MGMT-8741: Use full name when logging agent update
When checking if an agent image has changed, a full
image name was compared to an image tag. This
resulted in an agent update message being printed
to the log even when the image remained the same.

Comparing the full image name used by a registering
host to the full image name configured on the service
should solve the problem and produce an agent update
message (used for troubleshooting) only when there
was an actual change in agent image or version.