Backport podman installation mode changes (#3161)
* MGMT-8448: Use the image service when deploying with podman (#3023)

* MGMT-8448: Use the image service when deploying with podman

This deploys the image service as another container in the pod and
configures the assisted service to use it for downloading images.

https://issues.redhat.com/browse/MGMT-8448

* MGMT-8448: Redirect v1 image downloads to the image service

Users running the service using podman will be acustomed to downloading
the discovery image from the service directly. Since this is changing
with the addition of the image service we should add a redirect here
rather than letting those calls fail.

* NO-ISSUE: use images from edge-infrastructure org

Co-authored-by: Nick Carboni <ncarboni@redhat.com>