MGMT-8894: Simplify IPv6 Network Manager configuration
This commit simplifies confguration of the Network Manager in case IPv6
is used. Till now we were

* configuring consistent DHCPv6 client identification
* creating system-connections-merged overlay configuration
* manually configuring keyfile path

We are able to simplify this flow for the following reasons

* with the introduction of InfraEnvs, the static network configuration
  is now a property of InfraEnv and not of a cluster
* RHCOS handles creation of systemConnectionMerged whenever needed and
  if so, it handles configuration of the keyfile paths

Until now, the configuration applied by Assisted Service was not causing
problems as in RHCOS <=4.9 the keyfile paths were overriden whenever a
different overlay directory has been created.

Starting from RHCOS 4.10 this is no longer the case and network
configurations do not need to be stored in any merged directory. With
the removal of the custom keyfile path, we are leveraging the following
properties

* default path is /etc/NetworkManager/system-connections
* aforementioned path contains all the configurations coming from the
  custom NMStateConfig

Closes: MGMT-8894
Closes: Bug-2030289
Relates-to: Bug-2040195