MGMT-8716: uses Pod definition with podman
The Pod definition and ConfigMap definition can be retrieved by
themselves and used to run assisted-service locally without needing to
clone the repository or use developer tools like `make`. It also
presents a more familiar definition for the Pod than the several lines
of shell script it replaces.

Per the related issue
[MGMT-8715](https://issues.redhat.com/browse/MGMT-8715) it is no longer
required to mount in the nginx.conf file to the UI container, so it is
removed by this PR.