MGMT-8783: support multiple patch versions of os/release images
Added support to maintain OS and release images with cpu_architecture
in major.minor.patch format. This functionality is a means to allow
multiple z-stream versions concurrently, instead of the previous behaviour
which limited a single z-stream for every y-stream.

The following changes were made in versions package in order to
introduce the new behaviour:
* GetReleaseImgae:
 * Returns an image based on the specified x.y.z.
 * If missing, fallback to x.y (as previously).
* GetOsImage:
 * Returns an image based on the specified x.y.z.
 * If missing, fallback to x.y (as previously).
 * If x.y is not available, find latest applicable image for
   the specified x.y.z. This is done to avoid forcing a direct
   version match between OS and release image. E.g. for kube-api
   custom z-stream release versions support.