MGMT-8752: parsing error in 503 responses from router
The OCP router sends a 503 response with a standard HTML formatted message when the service is down.

However, the go client generated by swagger expects a JSON-formatted API message and fails to analyze the response body. The result is an unclear error message in the agent, installer, and controller when the service is down.

This PR replaces the 503 error message definition in the REST APIs with Infra Error construct that accommodates both external and application errors.