MGMT-8894: Catch FCs and nightlies when comparing OCP versions (#3226)
This commit extends the OCP version used when comparing in order to
comply with rules of the semantic versioning and release naming schema
in OCP.

This is because only "4.10" does not catch "4.10.0-fc.1" nor e.g. a
nightly release "4.10.0-0.nightly-2022-01-23-013716" - as per rules,
pre-releases have lower priority over normal versions.

In order to catch all of those, we are comparing against
"4.10.0-0.alpha" so that we cover FC releases as well as nightlies.

Closes: MGMT-8894
Closes: Bug-2030289
Relates-to: Bug-2040195