Bug 2035638: Stop BMH reconcile if InfraEnv label is missing
The reconcile loop was already stopping the BMH reconcile for CR's that
didn't have the infraenv label. However, this was happening after the
`inspect` label and cleaning were disabled, which caused BMAC to modify
BMH CR's that weren't part of the ZTP flow.

Move the infraenv check earlier in the `reconcileBMH` step so that BMAC
can stop the loop before doing changes to the CR.

Signed-off-by: Flavio Percoco <flavio@redhat.com>