MGMT-9035: Fix validation of CA bundle
This commit fixes the logic used to validate the CA bundle provided for
the purpose of oVirt platform.

Currently the validation passes when provided an incomplete bundle
because all the certificates were used as roots, including the
certificate that is a subject of the validation.

With this fix we are changing the behaviour so that all the certificates
in the bundle are identified as Root or Intermediary and afterwards the
verification is performed whether they can form a correct chain.

It is important to note that for a bundle to be considered valid every
certificate inside it must be correctly validating against a Root
certificate present in the aforementioned bundle.

Closes: MGMT-9035