NO-ISSUE: Use agent's status role instead of spec (#3187)
BMAC was using the agent spec's role field rather than the status. This worked fine because, so far, it's
been a requirement to set the role annotation for the BMH. However, if this annotation is not set, the service
will auto-dected the role of the host, which means the spec's role field will remain empty, while the status' one
will be correctly populated.

Since the status role field is always populated, BMAC should just stick to using that to know what role the agent is.

Signed-off-by: Flavio Percoco <flavio@redhat.com>

Co-authored-by: Flavio Percoco <flavio@redhat.com>