MGMT-8738: Set Disconnected status for media disconnection during discovery. (#3225)
Media disconnection is mostly caused by a poor network and can't be repaired without rebooting the machine, so we don't want the AI to perform steps on the host until the machine is rebooted.
Previously, if a discovery media was disconnected during the discovery phase, it would go into error status.  There is no way to return it to a non-error status without resetting the entire cluster.
Essentially, this PR enables recovery by rebooting the host machine.