MGMT-9057: Add KubeAPI support for schedulable masters property
SchedulableMasters property allows scheduling workload on the master
nodes of the OCP cluster. In #2075 the ability to set that has been
added to the REST API. In order to keep the feature parity with K8s API,
we are extending the AgentClusterInstall CRD so that we have ability to
use that feature.

Closes: MGMT-9057