MGMT-8744: Add log indicating that with auth none image url will not change (#3246)
When using authorisation types "local" and "RHSSO"; it is expected that the ISO image URL will change due to digital signatures added to the URL.
When using an authorisation type of "none" this is not expected. It is desired that we notify the user in the log that the URL will not be changing.
This change seeks to address that.