NO-ISSUE: Reduce cyclomatic complexity of ClusterDeployment updates
This commit reduces cyclomatic complexity of the updateIfNeeded()
function by extracting part responsible for updating Cluster, Machine
and Service networks.

As the Cluster object grows by adding more properties, this causes
development of the new logic more troublesome and requires refactoring
some of the already existing parts. This change here is one of those as
it does not impact the logic of the code and is purely a visual change.