MGMT-8047: Approve the CSRs after the installation (#3169)
Day1 CSRs are approved by assisted controller.  Since the controller is
not aware of day2 workers, until now these CSRs needed to be approved
manually.
This change applies on to ZTP flow, since in cloud assisted-service
cannot connect a cluster to attempt CSR approval.
The change is done in the agent controller that starts attempting to
approve CSRs when a day2 host moves to done.
At this point, the corresponding agent changes the stage to rebooting
which starts the CSR approval cycle.  It monitores the nodes in the
spoke cluster to find if a corresponding ready node is in the spoke
cluster.  If not, it keeps attempting to approve these CSRs.
Once a corresponding ready node is discovered, the stage in the agent is
moved to done and the cycle of CSR approval is stopped.