MGMT-9233: disk validation fails for day2 hosts (#3322)
The root cause of the problem was that a day2 cluster had an empty disk encryption configuration while a day2 host was created with an auto-assign role.

The empty disk encryption configuration caused the disk encryption validation to execute. At the same time, a mix of day1 and day2 logic in the validation's code failed the validation due to an absent role.

The fix is to separate day2 from day1's logic in the validation  (day2 logic does not need role assertion) and specifically assign a worker role to day2 hosts.

The latter will also prevent future problems stemming from validations that pends on role assignment.