MGMT-9134: convert image_test to v2
This is a pre-requisite code for removing V1 RegisterCluster.

It changes the flow of the image_test suite to use V2RegisterCluster.
As a follow-up, it also generates the ISO via InfraEnv as mixing of V1 and V2 flows is not possible.

The coverage rate may drop due to this PR. However, this is only a temporary drop until the main PR is submitted with some code deletion on the BE.