MGMT-9122: Remove DisableHost from subsystem
This is a pre-requisite code for removing V1 RegisterCluster.

EnableHost/DisableHost are APIs that were replaced in V2 by either unbinding to deregistration. Once we move to use V2 RegsiterCluster, the subsystem tests can no longer use those APIs.

This PR deletes the tests that aim to check EnableHost/DisableHost flow because we will soon remove those flows from the BE.

In other cases, DisableHost was used to induce an error. In these cases, the test generates another type of error instead or uses V2DergisterHost API to remove the host altogether.

V2UnbindHost API can not replace DisableHost because the InfraEnv in the tests is bound to a cluster. It is illegal to unbound a host from such an InfraEnv.

The coverage rate may be reduced by this PR, but this will be soon corrected by removing the DisableHost/EnableHost code from the BE.