OCPRHV-637: Allow disabling the scheduling on masters
Providers other than baremetal always set the compute
replicas to 0 in the 'install-config.yaml' file to
prevent the machine operator to provision additional
virtual machines at installation time (the compute
nodes installed by the Assisted Installer will be
`Nodes` only and not part of the `MachineSet`).

The side effect of this is that the generated manifests
will enable the scheduling of workloads in the control
plane (no compute nodes in the 'install-config.yaml').

So no matter how many nodes belong to the cluster the
scheduling of workloads on the control plane would
be enabled by default.

Closes: OCPRHV-637