MGMT-9321: Default to skipping tls verification in the operator
This changes the default behavior of the image service when deployed by
the operator to not verify certificates when downloading the template
images.

Most users of the operator will be mirroring the image content into
their environment so they will have to set this if they're not
configuring certs that are properly signed and trusted.

https://issues.redhat.com/browse/MGMT-9321