MGMT-9356: Add go mod to models (#3385)
it's a first step toward spliting kube-api api dir into a sub module.
kube-api crd definition is using definitions from the models directlry
to avoid cycle dependecy first need to split the model into a sub module
from the main go mod

reuqire specific gorm version because i had some issue with timezone
translation

Copy the modles go mod when building the container

go.mod is created once, if there are chnagese in the swagger file that
reuqire to update go.mod it will be done manually it's not common
process but if we call go mod every time we will have conflicts on the
pr checks every time package is updated event none of the code changes
so it's fair trade.