MGMT-9357 splint api dir into a sub module
It will enable to import api without involving in the main module replace
that we have in go.mod files.

Update golang version to 1.17 and update go.mod to golang 1.17
It will prevent some of the replaces that we need in go.mod file
Specifically resolving a conflict in sigs.k8s.io/json that uses
reflect package features from golang 1.17

Made go.mod to be compatible for both golang 1.16 and 1.17 it increase
the size of go.mod, it i selected to support only golang 1.17 go.mod
will size still increased so didn't found any issue supporting both

Fix code issues related to new k8s packages versions

Update hive and metal3 crds after version update

When generating manifests the command need to be executed from the api
directory, otherwise controller-gen will not set the sub modules and
will not generate anything even if directly passing ./api/... in the
paths parameter

k8s packages changed some of the behavior between the versions, looks
like right now update will return a new copy of the object
In our case Agent reconcile make Annotations and Labels maps if nil,
when adding a finalizer the code need to stop the flow and reconcile
again and not try to use the new object when the maps are nil again. It
caused panic in the service

After updating the k8s packages looks like admission-webhook server
don't support setting log level as parameter so i had to remove it when
running the webhooks server, didn't manage to find the reason, if it's
related to klog it looks like we are using the same version