MGMT-9206: Allow arm baremetal platform installation (#3345)
We want to allow baremetal platform installation on saas, for this we require  to override release image from which we will extract baremetal installer to the x86 one
This pr adds functionality to override arm image with x86 in case
ALLOW_INSTALLER_RELEASE_IMAGE_OVERRIDE env param is set to true, default
is false.
In case of arm and bm platform were choosed, we will get x86 release
image of the same version and extract openshift-baremetal-installer executer from it.