MGMT-8824: Copy additional NTP sources to chrony manifest.
When a user defined additional NTP sources (either via the API or via the
InfraEnv CR), the NTP sources already present on the host and these additional
NTP sources were expected to end up on the host's NTP sources, by applying the
chrony manifest created here to the hosts.

But when merging the host's NTP sources with the user added NTP sources, we took
into account only the sources with state "synchronised|synced". Which resulted
in only a subset of the merged sources to be present in the respective
[chrony](https://chrony.tuxfamily.org/) manifest and by consequence on the
host's NTP sources.

With this change, all user defined additional NTP sources, irrespective of their
state, are copied into the chrony manifest and are reflected in the chrony
sources.