MGMT-9527 fix generate-bundle after moving api to sub modules
operator-sdk doesn't know how to work with sub-modules so it require to
enter into api directory and run the commands from there, in addition
the PROJECT file need to be located in the same place, according to
operator-sdk code i saw that it is not configurable to set the path to
this file.
For some reason even after setting input and output dir it still creates
a local copy of config/manifests/kustomization.yaml so i had to do a
cleanup and remove it.