MGMT-8794: refresh hosts on cluster update (#3424)
When operators (CNV in particular) are enabled or disabled
the host validation status may change since the HW requirement
change as well.

The latter can cause a cluster to become not ready for installation.
However, since hosts are not refreshed synchronously, this is not relayed
immediately to the UI, causing a bad UX.

This PR refreshes the hosts within the UpdateCluster