MGMT-9539: Deduplicate existing presigned url models
Previously there were two url types defined in the swagger spec
`presigned` and `infra-env-image-url` both served the same purpose and
have a `url` attribute.

This commit combines these implementations into a new type,
`presigned-url` which will also be used for the upcoming endpoint to get
a url to download the ipxe boot script.

https://issues.redhat.com/browse/MGMT-9539