MGMT-9539: Allow image auth in query params for infra-env file downloads
This commit adds a new security definition `imageURLAuth` to handle
query parameter based auth for downloading infra-env files.

This is how a pre-signed URL for the ipxe-script will be implemented.

The authentication for this definition will be implemented using the
same functions as imageAuth. The only differences here are how the token
is provided (query param rather than header) and to which endpoints it
applies (only the infra-env files download for url-based).