MGMT-9349: Set signed boot artifact URLs in infra-env CR
These are only updated in the block when the ISO URL changes so that
they're not constantly being updated with new tokens.

This commit also extracts the url signing behavior for the kube-api case
into a separate common function so that it can be shared between the
events and boot artifacts code.