MGMT-9470 - Add a pin-latest.py script that pins catalog dependencies (#3468)
This script replaces occurrences of operand container image "URL"s in the
assisted operator's CSV manifest which have a `:latest` tag with "URL"s
that use a `@sha256:<hash>` tag instead. The hash chosen is simply the
hash of the image manifest that's currently behind `:latest` at the time
of running that script.

This script is going to be used by a work-in-progress prow job whose
purpose is generating an operator bundle image + catalog index image
containing said bundle image and pushing them to a public repository
in quay.io so it can be used by our test workflows and local
development workflows.