NO-ISSUE: generate support rootless podman
When doing things like `make generate-python-client`, we instructed
docker to map the local user into the container, so that the files will
have the right ownership after the Docker part of the build.

In podman, when running it rootless (there's no reason why we'd ever
need to run those steps with root podman), root inside the container is
already the right user outside of the container, so we need to prevent
this setting from applying.

The commit also makes sure that if you have the docker command supported
with podman, it will detect rightly that it is podman you are using.

Signed-off-by: Antoni Segura Puimedon <antoni@redhat.com>