MGMT-9277: Optimize connectivity check post step reply to improve load for large multi-node clusters (#3463)
Currently post step reply for connectivity check stores metrics for
packet loss and latency to all other hosts in the cluster.
In case there are 100 hosts in the cluster, 20000 metrics will be saved
(100 *100*2).
Since these metrics is not used, and saving them cause a lot of load,
they will be removed.