MGMT-9528: refactor code-generation scripts
This change aims to refactor how we generate code. Instead of spawning
docker containers for each of the tasks (generating go client,
generating Python client, generating backend-related files, etc.) it
will incorporate relevant tools/binaries from external images into the
assisted-service-build image so that it can be used in a prow job (that
cannot spawn docker-in-docker containers).

With those changes, we should be able to introduce a job named
verify-generated-code which makes sure that the developer of a PR has
generated all relevant code responding with the PR changes.