MGMT-9279: Remove assisted-service database timeout (#3355)
The on prem environment using `podman play kube` starts the service and
the database at the same time. This 10 second timeout causes the service
to give up on the database too quickly when the database is not quick
enough.

On actual kubernetes deployments this is not a serious issue because the
pod will simply crash loop and the situation will resolve itself.

On on-prem podman there's no automatic restarting, so the user has to
understand what happened and restart the service container manually,
which is inconvenient.

Removing this timeout as it's not really necessary - the service should
just keep trying to connect to the database